Modularized Overview codebase; images are rendering on the page as expected
